Move Q1 Rights API endpoints to internal controller (#619)
* Move Q1 Rights API endpoints to internal controller
#614
- Renamed existing RightsController to RightsInternalController and move all Q1 endpoints to it
- Added /internal/ path prefix to all endpoints
- Updated pats for all existing controller tests

* Updated Bruno manual test collection for new internal rights API paths

---------
