--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/Delegation/Authorized Parties Test Data Setup/Altinn App/Main to Person.bru
@@ -1,68 +1,0 @@
-meta {
-  name: Main to Person
-  type: http
-  seq: 3
-}
-
-post {
-  url: {{baseUrl}}/accessmanagement/api/v1/{{party}}/rights/delegation/offered
-  body: json
-  auth: none
-}
-
-headers {
-  Content-Type: application/json
-  Accept: application/json
-}
-
-body:json {
-  {
-      "to": [
-          {
-              "id": "urn:altinn:ssn",
-              "value": "{{toSsn}}"
-          },
-          {
-              "id": "urn:altinn:lastname",
-              "value": "{{toLastName}}"
-          }
-      ],
-      "rights": [
-          {
-              "resource": [
-                  {
-                      "id": "urn:altinn:org",
-                      "value": "{{org}}"
-                  },
-                  {
-                      "id": "urn:altinn:app",
-                      "value": "{{app}}"
-                  }
-              ],
-              "action": "read"
-          }
-      ]
-  }
-}
-
-vars:pre-request {
-  toSsn: '17105431031'
-  toLastName: Natvik
-  org: ttd
-  app: am-devtest-main-to-person
-  auth_tokenType: Personal
-  auth_userId: 20000542
-  auth_partyId: 50002650
-  auth_ssn: '08109412032'
-  party: 50006037
-}
-
-script:pre-request {
-  await tokenGenerator.getToken();
-}
-
-tests {
-  test("Should return 201 CREATED", function() {
-    expect(res.status).to.equal(201);
-  });
-}
