--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/Delegation/Org to Person/ttd-am-k6_badrequest.bru
@@ -1,62 +1,0 @@
-meta {
-  name: ttd-am-k6_badrequest
-  type: http
-  seq: 3
-}
-
-post {
-  url: {{baseUrl}}/accessmanagement/api/v1/{{party}}/rights/delegation/offered
-  body: json
-  auth: none
-}
-
-headers {
-  Content-Type: application/json
-  Accept: application/json
-}
-
-body:json {
-  {
-      "to": [
-          {
-              "id": "urn:altinn:ssn",
-              "value": "{{toSsn}}"
-          },
-          {
-              "id": "urn:altinn:lastname",
-              "value": "{{toLastName}}"
-          }
-      ],
-      "rights": [
-          {
-              "resource": [
-                  {
-                      "id": "urn:altinn:resource",
-                      "value": "{{resourceId}}"
-                  }
-              ],
-              "action": "ScopeAccess"
-          }
-      ]
-  }
-}
-
-vars:pre-request {
-  toSsn: '02056260016'
-  toLastName: Rimstad
-  resourceId: ttd-am-k6
-  auth_tokenType: Personal
-  auth_userId: 20000490
-  auth_partyId: 50002598
-  auth_ssn: '07124912037'
-}
-
-script:pre-request {
-  await tokenGenerator.getToken();
-}
-
-tests {
-  test("Should return 400 BADREQUEST", function() {
-    expect(res.status).to.equal(201);
-  });
-}
