--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/DelegationCheck/apps-test_success.bru
@@ -1,56 +1,0 @@
-meta {
-  name: apps-test_success
-  type: http
-  seq: 4
-}
-
-post {
-  url: {{baseUrl}}/accessmanagement/api/v1/{{party}}/rights/delegation/delegationcheck
-  body: json
-  auth: none
-}
-
-headers {
-  Content-Type: application/json
-  Accept: application/json
-}
-
-body:json {
-  {
-    "resource": [
-      {
-        "id": "urn:altinn:org",
-        "value": "{{org}}"
-      },
-      {
-        "id": "urn:altinn:app",
-        "value": "{{app}}"
-      }
-    ]
-  }
-}
-
-vars:pre-request {
-  org: ttd
-  party: 50005545
-  auth_tokenType: Personal
-  auth_userId: 20000490
-  auth_partyId: 50002598
-  auth_ssn: '07124912037'
-  app: apps-test
-}
-
-script:pre-request {
-  await tokenGenerator.getToken();
-}
-
-tests {
-  test("Should return 200 OK", function() {
-    expect(res.status).to.equal(200);
-  });
-  
-  test("Should return expected rights", function() {
-    const rights = res.getBody();
-    // ToDo Expect Result
-  });
-}
