--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/get -accessmanagement-api-v1--party--rights-delegation-offered.bru
@@ -1,30 +1,0 @@
-meta {
-  name: get -accessmanagement-api-v1--party--rights-delegation-offered
-  type: http
-  seq: 2
-}
-
-get {
-  url: http://accessmanagement/api/v1/{party}/rights/delegation/offered
-  body: json
-  auth: none
-}
-
-headers {
-  Authorization: {{apiKey}}
-}
-
-vars:pre-request {
-  resource: generic-access-resource
-  party: 50005545
-  auth_tokenType: Personal
-  auth_userId: 20000490
-  auth_partyId: 50002598
-  auth_ssn: '07124912037'
-}
-
-tests {
-  test("Should return 200 OK", function() {
-    expect(res.status).to.equal(200);
-  });
-}
