--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/post -accessmanagement-api-v1--party--rights-delegation-delegationcheck.bru
@@ -1,15 +1,0 @@
-meta {
-  name: post -accessmanagement-api-v1--party--rights-delegation-delegationcheck
-  type: http
-  seq: 1
-}
-
-post {
-  url: http://accessmanagement/api/v1/{party}/rights/delegation/delegationcheck
-  body: json
-  auth: none
-}
-
-headers {
-  Authorization: {{apiKey}}
-}
