--- test/Bruno/Altinn.AccessManagement/Manual Test Collection/Rights/post -accessmanagement-api-v1--party--rights-delegation-offered.bru
@@ -1,15 +1,0 @@
-meta {
-  name: post -accessmanagement-api-v1--party--rights-delegation-offered
-  type: http
-  seq: 2
-}
-
-post {
-  url: http://accessmanagement/api/v1/{party}/rights/delegation/offered
-  body: json
-  auth: none
-}
-
-headers {
-  Authorization: {{apiKey}}
-}
