+++ test/Bruno/Altinn.AccessManagement/Manual Test Collection/RightsInternal/CreateDelegation/Authorized Parties Test Data Setup/Generic Access Resource/Person to Person.bru
@@ -1,0 +1,63 @@
+meta {
+  name: Person to Person
+  type: http
+  seq: 5
+}
+
+post {
+  url: {{baseUrl}}/accessmanagement/api/v1/internal/{{party}}/rights/delegation/offered
+  body: json
+  auth: none
+}
+
+headers {
+  Content-Type: application/json
+  Accept: application/json
+}
+
+body:json {
+  {
+      "to": [
+          {
+              "id": "urn:altinn:ssn",
+              "value": "{{toSsn}}"
+          },
+          {
+              "id": "urn:altinn:lastname",
+              "value": "{{toLastName}}"
+          }
+      ],
+      "rights": [
+          {
+              "resource": [
+                  {
+                      "id": "urn:altinn:resource",
+                      "value": "{{resourceId}}"
+                  }
+              ],
+              "action": "read"
+          }
+      ]
+  }
+}
+
+vars:pre-request {
+  toSsn: '17105431031'
+  toLastName: Natvik
+  resourceId: devtest_gar_authparties-person-to-person
+  auth_tokenType: Personal
+  auth_userId: 20000542
+  auth_partyId: 50002650
+  auth_ssn: '08109412032'
+  party: 50002650
+}
+
+script:pre-request {
+  await tokenGenerator.getToken();
+}
+
+tests {
+  test("Should return 201 CREATED", function() {
+    expect(res.status).to.equal(201);
+  });
+}
