+++ test/Bruno/Altinn.AccessManagement/Manual Test Collection/RightsInternal/DelegationCheck/devtest_gar_bloop.bru
@@ -1,0 +1,52 @@
+meta {
+  name: devtest_gar_bloop
+  type: http
+  seq: 3
+}
+
+post {
+  url: {{baseUrl}}/accessmanagement/api/v1/internal/{{party}}/rights/delegation/delegationcheck
+  body: json
+  auth: none
+}
+
+headers {
+  Content-Type: application/json
+  Accept: application/json
+  Altinn-Party-OrganizationNumber: {{partyOrganizationNumber}}
+}
+
+body:json {
+  {
+    "resource": [
+      {
+        "id": "urn:altinn:resource",
+        "value": "{{resourceId}}"
+      }
+    ]
+  }
+}
+
+vars:pre-request {
+  resourceId: devtest_gar_bloop
+  party: 50005545
+  auth_tokenType: Personal
+  auth_userId: 20000490
+  auth_partyId: 50002598
+  auth_ssn: '07124912037'
+}
+
+script:pre-request {
+  await tokenGenerator.getToken();
+}
+
+tests {
+  test("Should return 200 OK", function() {
+    expect(res.status).to.equal(200);
+  });
+  
+  test("Should return expected rights", function() {
+    const rights = res.getBody();
+    // ToDo Expect Result
+  });
+}
