+++ test/Bruno/Altinn.AccessManagement/Manual Test Collection/RightsInternal/DelegationCheck/generic-access-resource_success.bru
@@ -1,0 +1,51 @@
+meta {
+  name: generic-access-resource_success
+  type: http
+  seq: 2
+}
+
+post {
+  url: {{baseUrl}}/accessmanagement/api/v1/internal/{{party}}/rights/delegation/delegationcheck
+  body: json
+  auth: none
+}
+
+headers {
+  Content-Type: application/json
+  Accept: application/json
+}
+
+body:json {
+  {
+    "resource": [
+      {
+        "id": "urn:altinn:resource",
+        "value": "{{resourceId}}"
+      }
+    ]
+  }
+}
+
+vars:pre-request {
+  resourceId: devtest-generic-access-resource
+  party: 50005545
+  auth_tokenType: Personal
+  auth_userId: 20000490
+  auth_partyId: 50002598
+  auth_ssn: '07124912037'
+}
+
+script:pre-request {
+  await tokenGenerator.getToken();
+}
+
+tests {
+  test("Should return 200 OK", function() {
+    expect(res.status).to.equal(200);
+  });
+  
+  test("Should return expected rights", function() {
+    const rights = res.getBody();
+    // ToDo Expect Result
+  });
+}
